spacefn: do not emit space if held for +250 msec with no other keypress
If you press space down with the intent of using the arrow keys to
navigate a page, say, but you change your mind and do not press any
other key, and let go of the space key even after a long while, space
is emitted. With many applications, space is mapped to "next page"...
so you're moving away from the current position whether you want it
or not.

With this change, holding the space key for long enough offers a way to
cancel the keypress. There is no particularly legitimate use for a long
keypress of a (non-repeating) key.